Fix file list
