+++ lib/al/Library/Collision/PartsConnectorUtil.cpp
@@ -11,7 +11,7 @@
 #include "Library/LiveActor/ActorPoseUtil.h"
 #include "Library/LiveActor/LiveActor.h"
 #include "Library/Math/MathUtil.h"
-#include "Library/Matrix/MatrixUtil.h"
+#include "Library/Math/MatrixUtil.h"
 #include "Library/Placement/PlacementFunction.h"
 #include "Library/Screen/ScreenPointTarget.h"
 #include "Project/HitSensor/HitSensor.h"
