+++ lib/al/Library/Math/MatrixUtil.cpp
@@ -1,4 +1,4 @@
-#include "Library/Matrix/MatrixUtil.h"
+#include "Library/Math/MatrixUtil.h"
 
 #include "Library/Math/MathUtil.h"
 
@@ -280,4 +280,23 @@ void preScaleMtx(sead::Matrix34f* outMtx, const sead::Vector3f& scale) {
     outMtx->scaleBases(scale.x, scale.y, scale.z);
 }
 
+const s32 bayerMatrix2[2][2] = {{0, 2}, {3, 1}};
+
+void makeBayerMatrix(s32* outMtx, s32 size) {
+    for (s32 y = 0; y < 1 << size; ++y) {
+        for (s32 x = 0; x < 1 << size; x++) {
+            s32 value = 0;
+            s32 shift = 0;
+            for (s32 k = size; k != 0 && size > 0; k--) {
+                s32 bitX = (x % (1 << k)) / (1 << (k - 1));
+                s32 bitY = (y % (1 << k)) / (1 << (k - 1));
+
+                value += bayerMatrix2[bitY][bitX] << shift;
+                shift += 2;
+            }
+
+            outMtx[x + (y << size)] = value;
+        }
+    }
+}
 }  // namespace al
