+++ src/Item/CoinStackBound.cpp
@@ -12,7 +12,7 @@
 #include "Library/LiveActor/ActorPoseUtil.h"
 #include "Library/LiveActor/ActorSensorUtil.h"
 #include "Library/Math/MathUtil.h"
-#include "Library/Matrix/MatrixUtil.h"
+#include "Library/Math/MatrixUtil.h"
 #include "Library/Movement/FlashingTimer.h"
 #include "Library/Nature/NatureUtil.h"
 #include "Library/Nature/WaterSurfaceFinder.h"
